meetup-8/dnasearch: Problem 5 test cases and trivial solution
You can just use Python's `s in t` syntax on strings to trivially do
this, though it's a) not what was asked for and b) we don't know the
actual algorithmic complexity of it (though it's almost certainly
faster than anything we might do in Python).

So we use it to validate the test cases, anyway.